Bug 35707: (alternate) Add clearfix to item statuses
This proposed alterantive patch adds changes the status <span>s to
<div>s. A div is basically the block-level equivalent of a span. This
has the same effect of pushing each status onto its own line.

The patch also adds generic "item_status" class to the <div>s as well so
that we can attach a consistent style to all statuses. The patch
includes a small amount of padding on these items in order to repeat
legibility.

To test, apply the patch and rebuild the staff interface CSS.

1) Add an item to stockrotation and trigger the first transfer
2) Place a hold on the item
3) Note that you will now see two statuses, one above the other as
   displayed in the screenshots attached to the bug.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>