Bug 35820: Fix table id on hold ratios page
This patch corrects the ID on the table showing holds on the holds
ratios page so that it matches the DataTables initialization and
configuration. I'm not aware of any bug with the functionality, but
consistency is good.

The patch also wraps the table in a "page-section" div, an update which
includes indentation changes.

To test, apply the patch and if necessary, place multiple holds on a
title so that the hold ratios page will display results.

- Go to Circulation -> Hold ratios.
- The table of hold ratios should look correct.
- Click the "Configure" button at the top of the table.
- On the table settings page, change the default visibility of some
  columns.
- Return to the hold ratios page to confirm that the correct columns are
  hidden by default.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>
(cherry picked from commit 05cdc48fd6c549f5f67af2562f9c5f889b63fa50)
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>